Code refactored as per review feedbacks
