+++ app/controllers/SessionsController.scala
@@ -256,7 +256,6 @@ class SessionsController @Inject()(val messagesApi: MessagesApi,
             Future.successful(BadRequest(views.html.sessions.updatesession(formWithErrors, formIds)))
           },
           sessionUpdateInfo => {
-            println("===============>" + sessionUpdateInfo)
             sessionsRepository
               .update(sessionUpdateInfo)
               .flatMap { result =>
