+++ app/views/feedback/feedbackformpreview.scala.html
@@ -1,5 +1,5 @@
 
-<div class="modal fade " id="myModal" role="dialog">
+<div class="modal fade" id="myModal" role="dialog">
     <div class="modal-dialog ">
 
         <div class="modal-content ">
@@ -32,4 +32,4 @@
     </div>
 </div>
 
-<script src="@routes.Assets.versioned("javascripts/feedbackFormPreview.js")" type="text/javascript"></script>
+<script src="@routes.Assets.versioned("javascripts/feedbackformpreview.js")" type="text/javascript"></script>
